+++ packages/zundo/src/testing/__tests__/zundo.test.ts
@@ -1,8 +1,9 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 vi.mock('zustand/vanilla');
-import { TemporalState, Write, zundo } from '../../index';
+import { zundo } from '../../index';
 import createVanilla, { StoreApi } from 'zustand/vanilla';
 import { act } from 'react-dom/test-utils';
+import { TemporalState, Write } from '../../types';
 
 interface MyState {
   count: number;
