+++ packages/zundo/src/types.ts
@@ -1,0 +1,14 @@
+import { TemporalStateWithInternals } from "./temporal";
+
+export type PopArgument<T extends (...a: never[]) => unknown> = T extends (
+  ...a: [...infer A, infer _]
+) => infer R
+  ? (...a: A) => R
+  : never;
+
+export type Write<T, U> = Omit<T, keyof U> & U;
+
+export type TemporalState<TState> = Omit<
+  TemporalStateWithInternals<TState>,
+  '__internal'
+>;
